+++ CHANGELOG.rst
@@ -5,6 +5,8 @@ dev
 - Fixed a Set Cookie Expires Days Bug
 - Implemented HTTP 451(Unavailable For Legal Reasons)
 - Print Exceptions in Protocol in debug mode
+- Deprecation Tools
+- Introduce FutureFinity Logging Facility
 
 v0.2.1
 ------